rf modal start

+++ src/shared/data/documents/user/user-support.ts
@@ -5,7 +5,7 @@ import { StructuredIdSchema } from '../learning-cycle/learning-cycle-support';
 export const ActiveLearningCycleSchema = StaticLearningCycleDataSchema.and(
   z.object({
     attemptingProblemStructuredIds: z.array(StructuredIdSchema),
-    actualTestDurationMs: z.number().int().min(0),
+    actualTestDurationMs: z.number().min(0),
     sessionStartedAt: z.number().int().min(0),
   })
 );
